arch-x86: Fix assertion due to clflush in DataTranslation::finish (#2772)
CLFLUSHOPT/WB/FLUSH are treated as reads to avoid page faults when done
on write-protected pages. This results in an assertion when
`translateTiming `calls `translation->finish` as it expects the `mode
`to be consistent with `state->mode`. To fix this keep original mode and
pass that to finish function

related PR: [#1080](https://github.com/gem5/gem5/pull/1080)